+++ SDK/c++/workbench.cpp
@@ -5,7 +5,7 @@
 #include <iostream>
 
 
-Workbench::Workbench(size_t type_id, double x0, double y0, size_t frame_remain, size_t materials_status,
-              size_t product_status) : type_id_(type_id), x0_(x0), y0_(y0), frame_remain_(frame_remain),
+Workbench::Workbench(int type_id, double x0, double y0, int frame_remain, int materials_status,
+              int product_status) : type_id_(type_id), x0_(x0), y0_(y0), frame_remain_(frame_remain),
               materials_status_(materials_status), product_status_(product_status){}
