+++ SDK/c++/workbench.h
@@ -11,11 +11,11 @@
 class Workbench {
 
 public:
-    size_t type_id_, frame_remain_, materials_status_, product_status_; // 工作台类型，剩余帧数，材料状态，产品状态。
+    int type_id_, frame_remain_, materials_status_, product_status_; // 工作台类型，剩余帧数，材料状态，产品状态。
     double x0_, y0_; // 坐标
 
-    Workbench(size_t type_id, double x0, double y0, size_t frame_remain, size_t materials_status,
-              size_t product_status);
+    Workbench(int type_id, double x0, double y0, int frame_remain, int materials_status,
+              int product_status);
 
     friend class Robot;
 };
